Update 04/01/24
update terbaru

+++ database/migrations/2023_12_26_113100_create_general_table.php
@@ -13,12 +13,12 @@ return new class extends Migration
     {
         Schema::create('general', function (Blueprint $table) {
             $table->unsignedBigInteger('general_id')->nullable();
-            $table->integer('general_karyawan_id');
-            $table->integer('general_nomor_kartu_akses');
+            $table->bigInteger('general_karyawan_id');
+            $table->bigInteger('general_nomor_kartu_akses');
             $table->string('general_firstname');
             $table->string('general_lastname');
             $table->string('general_nickname');
-            $table->integer('general_nik');
+            $table->bigInteger('general_nik');
             $table->string('general_tempat_lahir');
             $table->date('general_tanggal_lahir');
             $table->string('general_jenis_kelamin');
@@ -29,11 +29,11 @@ return new class extends Migration
             $table->string('general_golongan_darah');
             $table->string('general_alamat_ktp',500);
             $table->string('general_alamat_domisili',500);
-            $table->integer('general_npwp');
-            $table->integer('general_bpjs_ketenagakerjaan');
-            $table->integer('general_bpjs_kesehatan');
-            $table->integer('general_no_rek');
-            $table->integer('general_no_hp');
+            $table->bigInteger('general_npwp');
+            $table->bigInteger('general_bpjs_ketenagakerjaan');
+            $table->bigInteger('general_bpjs_kesehatan');
+            $table->bigInteger('general_no_rek');
+            $table->bigInteger('general_no_hp');
             $table->string('general_email');
             $table->string('general_medsos');
             $table->string('general_ukuran_baju',50);
@@ -44,10 +44,10 @@ return new class extends Migration
             $table->string('general_pendidikan_b');
             $table->string('general_keterampilan');
             $table->string('general_izin_tipe');
-            $table->integer('general_izin_nomor');
+            $table->bigInteger('general_izin_nomor');
             $table->date('general_izin_masa');
             $table->string('general_nama_kontak_darurat');
-            $table->integer('general_nohp_kontak_darurat');
+            $table->bigInteger('general_nohp_kontak_darurat');
             $table->string('general_lokasi_perusahaan');
             $table->string('general_lokasi_negara');
             $table->string('general_lokasi_site');
